Move HostConnection to the correct package name
