+++ app/src/main/java/org/xbmc/kore/service/library/SyncMusicVideos.java
@@ -22,7 +22,7 @@ import android.os.Bundle;
 import android.os.Handler;
 
 import org.xbmc.kore.jsonrpc.ApiCallback;
-import org.xbmc.kore.jsonrpc.HostConnection;
+import org.xbmc.kore.host.HostConnection;
 import org.xbmc.kore.jsonrpc.method.VideoLibrary;
 import org.xbmc.kore.jsonrpc.type.VideoType;
 import org.xbmc.kore.provider.MediaContract;
