+++ app/src/main/java/org/xbmc/kore/ui/sections/hosts/HostFragmentManualConfiguration.java
@@ -36,7 +36,7 @@ import org.xbmc.kore.eventclient.EventServerConnection;
 import org.xbmc.kore.host.HostInfo;
 import org.xbmc.kore.jsonrpc.ApiCallback;
 import org.xbmc.kore.jsonrpc.ApiException;
-import org.xbmc.kore.jsonrpc.HostConnection;
+import org.xbmc.kore.host.HostConnection;
 import org.xbmc.kore.jsonrpc.method.Application;
 import org.xbmc.kore.jsonrpc.method.JSONRPC;
 import org.xbmc.kore.jsonrpc.type.ApplicationType;
